ensure backup/restore naming standards in m365
Rename "data_collections" to "backup".  Ensure Backup and
Restore files mimic standard naming patterns, instead of using
arbitrary names for common behaviors.

No logic changes in this PR.  Only code movement/renaming.
