+++ src/internal/m365/backup.go
@@ -15,8 +15,6 @@ import (
 	"github.com/alcionai/corso/src/internal/m365/graph"
 	"github.com/alcionai/corso/src/internal/m365/onedrive"
 	"github.com/alcionai/corso/src/internal/m365/sharepoint"
-	"github.com/alcionai/corso/src/internal/m365/support"
-	"github.com/alcionai/corso/src/pkg/backup/details"
 	"github.com/alcionai/corso/src/pkg/control"
 	"github.com/alcionai/corso/src/pkg/fault"
 	"github.com/alcionai/corso/src/pkg/filters"
@@ -87,7 +85,7 @@ func (gc *GraphConnector) ProduceBackupCollections(
 
 	switch sels.Service {
 	case selectors.ServiceExchange:
-		colls, ssmb, canUsePreviousBackup, err = exchange.DataCollections(
+		colls, ssmb, canUsePreviousBackup, err = exchange.ProduceBackupCollections(
 			ctx,
 			gc.AC,
 			sels,
@@ -102,7 +100,7 @@ func (gc *GraphConnector) ProduceBackupCollections(
 		}
 
 	case selectors.ServiceOneDrive:
-		colls, ssmb, canUsePreviousBackup, err = onedrive.DataCollections(
+		colls, ssmb, canUsePreviousBackup, err = onedrive.ProduceBackupCollections(
 			ctx,
 			gc.AC,
 			sels,
@@ -118,7 +116,7 @@ func (gc *GraphConnector) ProduceBackupCollections(
 		}
 
 	case selectors.ServiceSharePoint:
-		colls, ssmb, canUsePreviousBackup, err = sharepoint.DataCollections(
+		colls, ssmb, canUsePreviousBackup, err = sharepoint.ProduceBackupCollections(
 			ctx,
 			gc.AC,
 			sels,
@@ -224,60 +222,4 @@ func checkServiceEnabled(
 	}
 
 	return true, canMakeDeltaQueries, nil
-}
-
-// ConsumeRestoreCollections restores data from the specified collections
-// into M365 using the GraphAPI.
-// SideEffect: gc.status is updated at the completion of operation
-func (gc *GraphConnector) ConsumeRestoreCollections(
-	ctx context.Context,
-	backupVersion int,
-	sels selectors.Selector,
-	restoreCfg control.RestoreConfig,
-	opts control.Options,
-	dcs []data.RestoreCollection,
-	errs *fault.Bus,
-) (*details.Details, error) {
-	ctx, end := diagnostics.Span(ctx, "connector:restore")
-	defer end()
-
-	ctx = graph.BindRateLimiterConfig(ctx, graph.LimiterCfg{Service: sels.PathService()})
-
-	var (
-		status *support.ConnectorOperationStatus
-		deets  = &details.Builder{}
-		err    error
-	)
-
-	switch sels.Service {
-	case selectors.ServiceExchange:
-		status, err = exchange.RestoreCollections(ctx, gc.AC, restoreCfg, dcs, deets, errs)
-	case selectors.ServiceOneDrive:
-		status, err = onedrive.RestoreCollections(
-			ctx,
-			onedrive.NewRestoreHandler(gc.AC),
-			backupVersion,
-			restoreCfg,
-			opts,
-			dcs,
-			deets,
-			errs)
-	case selectors.ServiceSharePoint:
-		status, err = sharepoint.RestoreCollections(
-			ctx,
-			backupVersion,
-			gc.AC,
-			restoreCfg,
-			opts,
-			dcs,
-			deets,
-			errs)
-	default:
-		err = clues.Wrap(clues.New(sels.Service.String()), "service not supported")
-	}
-
-	gc.incrementAwaitingMessages()
-	gc.UpdateStatus(status)
-
-	return deets.Details(), err
 }
