+++ src/internal/m365/exchange/collection.go
