+++ src/internal/m365/sharepoint/list_info.go
